src/main/java/com/kyoshii/service/CustomerService.java
package com.kyoshii.service;

import com.kyoshii.helper.DatabaseHelper;
import com.kyoshii.model.Customer;

import java.util.List;
import java.util.Map;

/**
 * @author xuzhiwei
 */
public class CustomerService {

    public List<Customer> getCustomerList() {
        String sql = "select * from customer";
        return DatabaseHelper.queryEntityList(Customer.class, sql);
    }

    public Customer getCustomer(long id) {
        String sql = "select * from customer where id=?";
        return DatabaseHelper.queryEntity(Customer.class, sql, id);
    }

    public boolean createCustomer(Map<String, Object> fieldMap) {
        return DatabaseHelper.insertEntity(Customer.class, fieldMap);
    }

    public boolean updateCustomer(long id, Map<String, Object> fieldMap) {
        return DatabaseHelper.updateEntity(Customer.class, id, fieldMap);
    }

    public boolean deleteCustomer(long id) {
        return DatabaseHelper.executeDelete(Customer.class, id);
    }
}
